Moved components into proper directories
