+++ services/web-app/components/gallery/galleryRow/index.tsx
@@ -1,0 +1,2 @@
+export * from "./rowTop";
+export * from "./rowBottom";
