+++ services/web-app/helpers/index.ts
@@ -1,6 +1,5 @@
 export * from "./erc1155";
 export * from "./erc721";
 export * from "./getNftMetaData";
-export * from "./getartistdata";
 export * from "./getipfs";
 export * from "./parseIpfs";
