+++ services/web-app/pages/_document.tsx
@@ -3,6 +3,7 @@ import { ServerStyleSheet } from 'styled-components';
 
 export default class MyDocument extends Document {
 
+  // This hydrates the pages with styled-components
   static async getInitialProps(ctx: DocumentContext) {
     const sheet = new ServerStyleSheet()
     const originalRenderPage = ctx.renderPage
